+++ src/radical/analytics/__init__.py
@@ -12,6 +12,13 @@ from .plotter    import Plotter
 
 # ------------------------------------------------------------------------------
 #
+from .utils import get_plotsize, get_mplstyle, stack_transitions
+from .utils import get_pilot_series, get_plot_utilization, get_pilots_zeros
+from .utils import to_latex
+
+
+# ------------------------------------------------------------------------------
+#
 import os
 import radical.utils as ru
 
@@ -24,6 +31,13 @@ version = version_short
 logger = ru.Logger('radical.analytics')
 logger.info('radical.analytics    version: %s' % version_detail)
 
+
+# ------------------------------------------------------------------------------
+#
+__all__ = ('Experiment','Session','Entity','Plotter', 'get_plotsize',
+           'get_mplstyle', 'stack_transitions', 'get_pilot_series',
+           'get_plot_utilization', 'get_pilots_zeros', 'to_latex')
+
+
 # ------------------------------------------------------------------------------
 
-__all__ = ('Experiment','Session','Entity','Plotter',)